server/requirements.txt
django

djangorestframework
django-filter
django-colorfield
django-cors-headers

drf-spectacular

django-environ
django-storages[google]

psycopg2-binary

gunicorn

google-api-python-client
google-auth

pyyaml
model_bakery
Faker
httpx